--- src/app/game/layout.tsx
@@ -1,22 +1,0 @@
-import React from "react";
-
-export default function layout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <div className="w-full h-full overflow-x-hidden overflow-y-scroll">
-      <div className="fixed w-full h-full overflow-hidden z-0">
-        <div
-          className="bg-cover bg-center w-full h-full fixed z-[-1]"
-          style={{ backgroundImage: "url('/images/bg-game.jpg')" }}
-        ></div>
-      </div>
-      <div className="max-w-xl w-xl h-screen relative z-1 m-auto flex flex-col items-center justify-between gap-6">
-        <main className="flex-1 w-full h-full">{children}</main>
-        <footer className="w-full flex justify-center"></footer>
-      </div>
-    </div>
-  );
-}
