+++ src/app/game/new-stockmon/layout.tsx
@@ -1,0 +1,10 @@
+import CommonLayout from "@/components/ui/CommonLayout";
+import React from "react";
+
+export default function layout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return <CommonLayout>{children}</CommonLayout>;
+}
