+++ src/components/game/style.scss
@@ -51,7 +51,6 @@ button {
     left: 0;
     top: -25%;
     z-index: 5;
-    background-image: url("/icons/game-target.svg");
   }
   &__aim::after {
     background: url("/icons/game-target.svg") no-repeat 0px 0px;
@@ -61,7 +60,7 @@ button {
     width: 62px;
     background-size: 80%;
     top: -38px;
-    left: -24px;
+    left: -22px;
     display: inline-block;
   }
 
@@ -101,7 +100,9 @@ button {
 }
 
 .bounce-in-up {
-  animation: bounce-in-up 1s ease;
+  opacity: 0;
+  animation: bounce-in-up 1s 1.25s ease;
+  animation-fill-mode: forwards;
 }
 
 @keyframes bounce-in-up {
@@ -118,6 +119,7 @@ button {
   }
   100% {
     transform: scale(1) translateY(0);
+    opacity: 1;
   }
 }
 
@@ -135,5 +137,79 @@ button {
   }
   100% {
     transform: translateX(0);
+  }
+}
+
+.rotate-in-down-left {
+  animation: rotate-in-down-left 2s ease infinite;
+}
+@keyframes rotate-in-down-left {
+  0% {
+    transform-origin: left bottom;
+    transform: rotate(-60deg);
+    opacity: 0;
+  }
+  50% {
+    opacity: 1;
+  }
+  100% {
+    transform-origin: left bottom;
+    transform: rotate(0) translateY(-400px);
+    opacity: 1;
+  }
+}
+
+.catch {
+  position: absolute;
+  bottom: -4vw;
+  // border-radius: 100%;
+  height: 100%;
+  left: 1%;
+  animation: throw 1.05s 0.05s 1;
+  transform: translate3d(0, 0, 0) scale(1.5);
+  animation-fill-mode: forwards;
+  &.Miss {
+    animation: throw-fail 1.05s 0.05s 1;
+    animation-fill-mode: forwards;
+  }
+}
+
+@keyframes throw {
+  0% {
+    bottom: -40vh;
+  }
+  50% {
+    bottom: 13vh;
+  }
+
+  100% {
+    transform: translate3d(15vw, 5vh, -50vh) scale(0.75);
+  }
+}
+
+@function custom-random() {
+  // 두 범위 중 하나를 선택하기 위한 무작위 결정
+  $range: random(2);
+
+  // 첫 번째 범위: -150 ~ 5 (5 제외)
+  @if $range == 1 {
+    @return random(155) - 150; // 0~154 -> -150 ~ 4
+  }
+  // 두 번째 범위: 15 ~ 200 (15 포함)
+  @else {
+    @return random(186) + 15; // 0~185 -> 15 ~ 200
+  }
+}
+
+@keyframes throw-fail {
+  0% {
+    bottom: -40vh;
+  }
+  50% {
+    bottom: 13vh;
+  }
+
+  100% {
+    transform: translate3d(custom-random() + vw, 5vh, -50vh) scale(0.75);
   }
 }
